Keep user cache PRF keys in sync with encrypted envelope
Before: after adding a new passkey, the passkey cannot be used to log
in with a cached user, because the passkey's credential ID is not
listed in the cached user's PRF keys.

After: after adding a new passkey or deleting one, the PRF keys in the
corresponding cached user are updated automatically.