Restructure pageSettings.loggedPasskey and .manageOtherPasskey translation keys
The current grouping doesn't quite match how the keys are used: most of the
`loggedPasskey` entries are used for both the logged-in passkey and the list of
other passkeys, and most of the `manageOtherPasskey` entries are used only in
the new passkey registration procedure.

This restructures the translation keys to better reflect where they are used:
`passkeyItem` for the keys used in `<WebauthnCredentialItem>` both for the
logged-in passkey and the others, `registerPasskey` for the keys specific to the
registration procedure, and `title` for the section headings that don't fit
either of the previous two categories.