Protect private data update with ETag/If-Match headers
This helps prevent data loss if a user has two concurrent sessions and one
attempts to overwrite changes made by the other. The client now keeps track of
the `X-Private-Data-ETag` value returned when the client most recently updated
its local copy of the private data. Whenever a backend request is made, we check
if the request failed with status code 412 (Precondition Failed) and a
`X-Private-Data-ETag` response header. If that response header matches the
current local copy, then the request is retried with the
`X-Private-Data-If-Match` header set to this value. If this succeeds, the local
copy of `X-Private-Data-ETag` is updated to the value of the
`X-Private-Data-ETag` response header returned by the successful retried
request.

The local copy of `X-Private-Data-ETag` must be initialized explicitly using the
`updatePrivateDataEtag` function. This is done in the signup and login
functions.

These headers are meant to imitate the general-purpose `ETag` and `If-Match`
headers, but named with a prefix to indicate that they only apply to the private
data field when used in requests that also retrieve or handle other fields.