Add session timeout
We use `sessionStorage` for the session key in the hopes of bounding the
lifetime of the decrypted wallet contents, as `sessionStorage` is cleared when
the browser tab is closed. We chose this passive approach for fail-safety, so
that we don't rely on active application code to perform the cleanup - which
risks that the cleanup might be skipped in case the session ends unexpectedly,
like if the browser crashes or in case of power loss.

However, a session may continue indefinitely, so there is no guarantee that the
`sessionStorage` will ever be cleared. But none of this prevents us from using
active safety as well: either the session ends within some given time, in which
case the browser clears the `sessionStorage` and thus the decrypted wallet
contents; or the session does _not_ end within that time, in which case the
application is still running and can trigger the cleanup programmatically.

This sets a configurable session timeout with a default of 15 minutes. When the
timeout triggers, the key store is closed and the user needs to log in again in
order to access the wallet contents. The timer resets whenever the user
interacts with the application.