Fix redirection loop: extract SessionContext component
Commit 04b87a41f4516946d46875e515e7e4b2c03768b5 introduced a possible
redirection loop, because the redirect condition differs between
`<PrivateRoute>` and `<Login>` - `<PrivateRoute>` queries both `api` and
`LocalStorageKeystore` for the logged-in status, while `<Login>` queries only
`api`. Because of this, a redirection loop occurs if `/` is loaded while
`sessionState` is set in session storage but the keystore is not open:

```
Warning: Maximum update depth exceeded. This can happen when a component calls setState inside useEffect, but useEffect either doesn't have a dependency array, or one of the dependencies changes on every render.
  at Navigate
  at PrivateRoute
  at RenderedRoute
  at Routes
  at div
  at HandlerNotification
  at Suspense
  at Router
  at BrowserRouter
  at I18nextProvider
  at App
```

This extracts a common `SessionContext` to keep track of the logged-in state, so
that the same logic can be shared everywhere an "is logged in" check is needed.