Remove single-value .map in History.js
This `.map` causes the rendered result to be wrapped in an array, which causes
React to require a key for the fragment. This shows as this lint when opening
details of a credential:

```
Warning: Each child in a list should have a unique "key" prop.
```

Since always `.map`s over an array that is guaranteed to contain exactly one
element, there's no actual need for a React key if we just unwrap the array.