Save and recall WebAuthn credential transports into and from PRF key info
This will only take effect for newly created credentials, because the
`getTransports` function only exists on `navigator.credentials.create()`
results.

We do also have these transports stored in the backend and _could_ use that to
retroactively set the transports in the PRF key info in the encrypted container.
That seems like more work than it's worth, but could be revisited as an option
in the future.