refactor: extract searchable list component
Create a component to accept a list of elements and display them as buttons. The list is searchable by the intended field, with highlight matching characters. Also, there is now a new util function to help calling nested properties of an object, when the field has to be passed as a parameter.