feat: add language selector in UI
Refactor the language selector in order to display a pre-styled popup/modal with the available languages and add to the initial login page and the sidebar. The refactor has to do only with the appearance that could fit well in both scenarios, as long as it is a reusable component.