Fix corruption data race in cached users
The cause of this data corruption is: after successfully logging in and
unlocking the keystore, the `useEffect` in `LocalStorageKeystore` calls
`setCachedUsers` to update the cached users (to copy the `prfKeys` from the
logged-in user's `privateData` to the matching cached user). But when there are
two tabs open simultaneously, that code runs simultaneously in both tabs, and
the `globalUserHandleB64u` is different between the two tabs. The user1 tab has
`globalUserHandleB64u` set to the user handle of user1, and the user2 tab has
`globalUserHandleB64u` set to the user handle of user2. So the result is that
both cached users get updated with user2's `prfKeys`, which then causes the user
to be logged into user2 even if they press the cache button labeled "Log in as
user1".