Make Pushed Authorization Request optional as per spec
The spec does not require the Pushed Authorization Request (PAR)
endpoint to be in the authorization server (AS) metadata, yet the
wwwallet crashes if this pushed_authorzation_server_endpoint isn't
there.

This change makes it optional. It then initiates either a "normal" oidc
flow, if the AS doesn't support the PAR. But intitiates a PAR if the
server requires it.

This fixes issue #598.