Add wallet-common as submodule
This enables each commit in wallet-frontend to track a particular commit in
wallet-common, and enables easily checking out that version using `git submodule
update`. This is very helpful when moving between branches or running earlier
versions of wallet-frontend that may no longer be compatible with the up-to-date
main version of wallet-common.