Refactor foldAllEventsIntoBaseState: WalletStateContainer to foldState: WalletState
The current name and function signature
`foldAllEventsIntoBaseState(walletStateContainer: WalletStateContainer):
WalletStateContainer` is misleading, since this function does not empty the
`events` or update the `lastEventHash` of the returned container. Thus, the
returned container is not a valid `WalletStateContainer` since adding any events
to it will make the event history inconsistent.

Fortunately this function is currently used only to compute the effective state,
and not to add any new events to the returned container. Thus we can refactor
the function to return only the state component instead of the whole container.