Make history divergence point search more robust to corner cases
Previously, `findDivergencePoint` and `mergeEventHistories` had a couple of
issues with various edge cases:

- They assume that both event histories start from the same base state, and
  merge events incorrectly if they do not.
- `findDivergencePoint` assumes that at least one history branch has at least
  one unique event, and returns a divergence point that is too early if that is
  not the case.
- `findDivergencePoint` may throw an array index error if the first event in
  either list is determined to be the parent of the divergence point.

These changes expand the logic for finding the divergence point to handle these
and other edge cases more robustly. Hopefully it should now be safe to call
`mergeEventHistories` even with the same identical container as both arguments.