Don't require unlock to manage passkeys
This was originally necessary because we re-encrypted the key store to a session
key on login and didn't keep the main key unwrapped in the session, therefore we
needed to re-authenticate the user in order to unwrap the main key so it can be
re-wrapped under the new PRF key.

Now we do keep the main key unwrapped in the session, so we no longer need to
re-authenticate the user.

This still keeps the "unlock" requirement for deleting the account. This is
purely a client-side safeguard against accidents, and can be easily bypassed.