Make state and reducer types more correct
Version 3 of `WalletStateContainer` now allows both the base state and the
events to be version 3 _or earlier_. Similarly, reducers like `foldState`,
`foldNextEvent` and `walletStateReducer` now accept a mix of earlier schema
versions in the state and events, and return a state with uncertain schema
version. This is more accurate to what state containers may look like in
practice - since the base state and pre-existing events may be of any earlier
schema version - and also more flexible, since the input and output version no
longer need to exactly match. This is most apparent in tests of parsing and
building on pre-existing states from earlier versions, which previously would
easily lead to various type errors.